fix: mmseqs_search
- further structure simplification
